[MOV] mail: rename RTC models (step 2)
- rename 'RtcCallViewer' to 'CallView' (files)
- rename 'RtcActivityNotice' to 'CallSystrayMenu' (files)
- rename 'RtcCallParticipantCard' to 'CallParticipantCard' (files)
- rename 'RtcLayoutMenu' to 'CallLayoutMenu' (files)
- rename 'RtcConfigurationMenu' to 'CallSettingsMenu' (files)

Task-2868876
