+++ addons/mail/static/src/components/call_layout_menu/call_layout_menu.xml
