+++ addons/mail/static/src/components/call_participant_card/call_participant_card.scss
