+++ addons/mail/static/src/components/call_settings_menu/call_settings_menu.xml
