+++ addons/mail/static/src/components/call_settings_menu_device/call_settings_menu_device.xml
