+++ addons/mail/static/src/components/call_systray_menu/call_systray_menu.scss
