+++ addons/mail/static/src/components/call_systray_menu_container/call_systray_menu_container.js
@@ -2,7 +2,7 @@
 
 import { useModels } from '@mail/component_hooks/use_models';
 // ensure components are registered beforehand.
-import '@mail/components/rtc_activity_notice/rtc_activity_notice';
+import '@mail/components/call_systray_menu/call_systray_menu';
 import { getMessagingComponent } from "@mail/utils/messaging_component";
 
 const { Component } = owl;
