+++ addons/mail/static/src/components/call_systray_menu_container/call_systray_menu_container.xml
