+++ addons/mail/static/src/components/call_view/call_view.scss
