+++ addons/mail/static/src/models/call_layout_menu.js
