+++ addons/mail/static/src/models/call_participant_card.js
