+++ addons/mail/static/src/models/call_settings_menu.js
