+++ addons/mail/static/src/models/call_systray_menu.js
