+++ addons/mail/static/src/models/call_view.js
